requirements.txt
openai==1.55.3
tavily-python==0.3.3
python-dotenv==1.0.0
requests==2.31.0
APScheduler==3.10.4
beautifulsoup4==4.12.3
lxml==5.1.0
pymongo==4.16.0